all tpetra changes in one commit
- renames MeshFrameworkAlgorithms to MeshAlgorithms, many changes to get src/mesh compiling with Mesh=MeshCache<MemSpace_kind::DEVICE>
- some frameworks do not provide some things, must try/catch or ask for permission somehow

+++ src/solvers/SolverNox.hh
@@ -130,7 +130,7 @@ class SolverNox : public Solver<VectorClass, VectorSpace> {
 
   // mutators
   void set_tolerance(double tol) { tol_ = tol; }
-  void set_pc_lag(int pc_lag) { pc_lag_ = pc_lag; }
+  void set_pc_lag(double pc_lag) { pc_lag_ = pc_lag; }
   void set_db(const Teuchos::RCP<ResidualDebugger>& db) { db_ = db; }
 
   // access
@@ -150,7 +150,7 @@ class SolverNox : public Solver<VectorClass, VectorSpace> {
   Teuchos::RCP<ResidualDebugger> db_;
   Teuchos::RCP<SolverFnBaseJF<VectorClass, VectorSpace>> jf_fnbase_;
   double tol_;
-  int pc_lag_;
+  double pc_lag_;
 };
 
 } // namespace AmanziSolvers
